Possibilità di commentare e votare e alert nel caso non lo si possa fare(ospiti e amministatori-giornalisti) rimosso trigger gestito completamente java rimossi inserimenti voti dal file sql per via del trigger Da risolvere: bug per via dell'entità nel context

+++ src/main/java/model/dao/GiocoDAO.java
@@ -34,7 +34,7 @@ public class GiocoDAO {
                     rs.getString(1),
                     rs.getString(3),
                     rs.getInt(5),
-                    rs.getInt(4),
+                    rs.getFloat(4),
                     rs.getDate(2),
                     rs.getString(6),
                     new PiattaformaDAO(con).doRetriveByGame(titolo),
